requirements.txt
PySide6
SQLAlchemy
psycopg2-binary